🎨 chore: add new files and update existing files This commit adds new files and updates existing files. The new files include an image file for the visit calendar and an image file for the visit QR code. The existing files that were updated include the `StringConstant` class, which now has additional constants for the `Add Visit` page. The `StyleConstant` class was also added, which contains a `formLabel` constant. The `VisitEntities` class was added, which contains a list of `VisitEntities` objects. The `main.dart` file was updated to use the `AddVisitPage` instead of the `VisitPage`. The `AddActivityPage` and `AddActivityForm` files were also updated to use a time picker instead of a date picker for the start and end times.
✨ feat(add_visit_ui.dart): add AddVisitPage UI
This commit adds the UI for the AddVisitPage, which includes a form for adding a visit. The form includes fields for selecting a destination, entering a location, and selecting a date. It also includes a QR code and a button for submitting the form.

🎨 style(visit_ui_card.dart): fix constructor super call and add missing import
✨ feat(visit_ui_card.dart): create VisitCard widget
The VisitCard widget is created to display a visit card with the visit's date, title, and address. The widget is used in the VisitPage to display a list of visit cards. The FleetimeDatePickerReuseable widget is renamed to FleetimeTimePickerReuseable to better reflect its purpose. The FleetimeTextFieldReuseable widget's border is updated to use an outline border with rounded corners. The intl package is added as a dependency to support date formatting.