🚀 feat(AndroidManifest.xml, string_constant.dart, main.dart, add_activity_ui.dart, add_overtime_ui.dart): add overtime feature The changes include adding support for overtime feature in the application. The AndroidManifest.xml file now has an attribute to enable on back invoked callback. The string_constant.dart file now has constants for the add overtime page. The main.dart file now has the AddOvertimePage as the home page. The add_activity_ui.dart file has an unused field removed. The add_overtime_ui.dart file is a new file that contains the UI for the add overtime page.
🔨 refactor(add_visit_ui.dart): extract form to a separate widget
The form in the AddVisitPage is now extracted to a separate widget called AddVisitForm. This improves the readability and maintainability of the code. The AddVisitForm widget is now used in the AddVisitPage. The form submission logic is also updated to print the form values to the console when the submit button is pressed.

✨ feat(add_visit_ui_add_form.dart): add form for adding visit
✨ feat(form_reuseable.dart): add dropdown and datepicker form fields
This commit adds a new form for adding a visit in the `add_visit_ui_add_form.dart` file. The form contains dropdown and datepicker form fields that are implemented in the `form_reuseable.dart` file. The dropdown and datepicker form fields are reusable widgets that can be used in other forms.