🚀 feat(string_constant.dart): add constants for add permission page ✨ feat(main.dart): change home page to AbsentMapPage Added constants for the Add Permission page to improve code readability and maintainability. The AbsentMapPage is now set as the home page instead of the SuccessPage. This change is made to reflect the new feature that has been added to the application.
✨ feat(absent_maps_ui.dart): add AbsentMapPage and FleetimeCustomListTile widgets
This commit adds the AbsentMapPage widget, which displays a map using the FlutterMap package and a custom list tile widget called FleetimeCustomListTile. The FleetimeCustomListTile widget is used to display information about the user's presence status and allows the user to toggle their presence status using a switch. The AbsentMapPage widget also includes a floating container for displaying additional details about the user's presence status.

✨ feat(add_permission_ui.dart): add AddPermissionPage widget
✨ feat(pubspec.yaml): add latlong2 package dependency
The AddPermissionPage widget is added to the application. It contains a form for adding a permission type, date, photo, and description. The latlong2 package is added as a dependency to the pubspec.yaml file. This package provides a set of tools for working with latitude and longitude coordinates.