✨ feat(string_constant.dart): add constant values for presence data page and presence history page This commit adds constant values for the presence data page and presence history page. The constant values are used to display the title, summary, and other information on the pages. This improves the consistency of the application and makes it easier to maintain.
✨ feat(routes.dart): add route for presence history page
This commit adds a route for the presence history page. The route is added to the routes map, which is used to navigate between pages in the application.

✨ feat(presence_card_entieties.dart): add entities for presence data page
This commit adds entities for the presence data page. The entities are used to display the title and subtitle of the cards on the page.

🔥 chore(main.dart): remove LoginPage and add HomePage as the initial page
This commit removes the LoginPage and adds the HomePage as the initial page. This improves the user experience as

🎨 style(presence_ui.dart): format code and add imports
✨ feat(presence_ui.dart): add presence card list view
The code has been formatted to improve readability. The imports for the color and string constants have been added. A presence card list view has been added to the presence page. The list view displays a list of presence cards with their respective titles, subtitles, and icons.

🆕 feat(presence_history_ui.dart): add presence history page
A new file, presence_history_ui.dart, has been added to the project. The file contains the implementation for the presence history page. The page displays a tab view with two tabs, one for the daily presence history and another for the monthly presence history.