🆕 feat(shift-recap): add shift recap page and entities The shift recap page is added to the application, which displays a list of shift recaps. The shift recap entities are also added to the application, which contains the date, status, and shift time of each shift recap.
🎨 style(shift_recap_ui.dart): refactor ShiftRecapPage widget
The ShiftRecapPage widget has been refactored to include a TabBar and TabBarView to display two different views. The TabBar and TabBarView are controlled by a TabController. The AppBar has been updated to include a back button and an add button. The ListView has been updated to display a list of shiftRecapEntitiesList. The Card widget has been updated to display the date and status of each shift. The status is displayed as a Chip widget.