✨ feat(routes.dart): add new routes for various pages Added new routes for various pages such as presence recap, permission recap, overtime recap, salary history, salary history detail, shift recap, activity recap, and reimbursement recap. This allows the user to navigate to these pages directly from the app's home screen.
🐛 fix(presence_card_entieties.dart): add routeName property to PresenceCardEntities
Added a new property called routeName to the PresenceCardEntities class. This property is used to store the route name for each presence card entity. This allows the user to navigate to the corresponding page when they tap on a presence card.

🐛 fix(main.dart): change home page to LoginPage
Changed the home page to LoginPage instead of PresenceRecapPage. This is because the user should be directed to the login page first before they can access the app's features.

✨ feat(presence_ui.dart): add navigation to corresponding page when presence