✨ feat(routes.dart): add new route for AddVisitNewPage 🎨 style(add_visit_ui.dart): change title to 'Visit Kerja Lama' 🎨 style(add_visit_ui_new.dart): change title to 'Visit Kerja Baru' The new route for AddVisitNewPage is added to the routes map in routes.dart. The title of AddVisitPage is changed to 'Visit Kerja Lama' to reflect the fact that it is the old version of the page. The title of AddVisitNewPage is changed to 'Visit Kerja Baru' to reflect the fact that it is the new version of the page.
🎨 style(add_visit_ui_add_form.dart): add spacing and dropdown for client selection
Added a dropdown for client selection and added spacing for better UI readability.

🚀 feat(visit_ui.dart): add support for adding new and existing clients
Added support for adding new and existing clients. If the tab controller is on index 0, the user is directed to add a new visit for a new client. If the tab controller is on index 1, the user is directed to add a new visit for an existing client.